+++ flutter/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.1.0
 
 - Add error messages
+- Rewrite example
+- Create gh-pages with Flutter Demo
+- Implement Widget update
 
 ## 1.0.0